+++ arithmetic/subtraction.go
@@ -1,7 +1,7 @@
 // arithmetic/subtraction.go
 package arithmetic
 
-// Subtract returns the difference between two integers.
-func Subtract(a, b int) int {
+// Subtract returns the difference between two float64 numbers.
+func Subtract(a, b float64) float64 {
     return a - b
 }
